Refactor API imports and implement proxy API service for semester exam clearance and user authentication
