+++ Not-Student-Portal-v2/src/app/dashboard/page.tsx
@@ -15,7 +15,7 @@ import {
   SGPAData,
   profileService,
   StudentInfo
-} from '@/services/api';
+} from '@/services/proxy-api';
 
 export default function DashboardPage() {
   const pageTitle = 'Student Dashboard';
