+++ Not-Student-Portal-v2/src/components/dashboard/StatCards.tsx
@@ -8,7 +8,7 @@ import {
   dashboardService, 
   calculatePaymentSummary, 
   PaymentSummary 
-} from "@/services/api";
+} from "@/services/proxy-api";
 
 // Props Interface
 interface StatCardsProps {
