+++ Not-Student-Portal-v2/src/components/forgot-password/components/reset-password.tsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
-import { authService } from "../../../services/api";
+import { authService } from "../../../services/proxy-api";
 import { Button } from "../../ui/button";
 import {
   Form,
