+++ Not-Student-Portal-v2/src/components/forgot-password/components/verify-code.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
-import { authService } from "../../../services/api";
+import { authService } from "../../../services/proxy-api";
 import { Button } from "../../ui/button";
 import {
   Form,
