+++ Not-Student-Portal-v2/src/components/profile-tabs/GuardianInfoTab.tsx
@@ -2,7 +2,7 @@
 import React, { memo } from 'react';
 import { Card, CardHeader, CardContent } from '@/components/ui/card';
 import { Table, TableBody, TableCell, TableRow } from '@/components/ui/table';
-import { StudentInfo } from '@/services/api';
+import { StudentInfo } from '@/services/proxy-api';
 
 interface GuardianInfoTabProps {
   studentInfo: StudentInfo | null;
