+++ Not-Student-Portal-v2/src/contexts/AuthContext.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React, { createContext, useContext, useState, useEffect } from 'react';
-import { LoginResponse } from '@/services/api';
+import { LoginResponse } from '@/services/proxy-api';
 
 interface AuthContextType {
   user: LoginResponse | null;
